refactor(tests): split work-task-service.test.ts into 8 focused files
The 1926-line monolith is replaced by 8 test files under 600 lines each,
plus a shared helpers module. No tests were removed or changed — just split
by logical concern (creation, bun-install, lifecycle, validation, repo-map,
PR dedup, shutdown, governance).
